+++ models/eprop_synapse.h
@@ -396,9 +396,9 @@ eprop_synapse< targetidentifierT >::send( Event& e, size_t thread, const EpropCo
 
     if ( t_spike >= t_next_update_ + shift )
     {
-      int idx_current_update = static_cast< int >( std::floor( ( t_spike - dt ) / update_interval ) );
+      long idx_current_update = static_cast< long >( std::floor( ( t_spike - dt ) / update_interval ) );
       double t_current_update_ = idx_current_update * update_interval;
-      int current_optimization_step_ = 1 + idx_current_update / cp.batch_size_;
+      long current_optimization_step_ = 1 + idx_current_update / cp.batch_size_;
 
       target->write_update_to_history( t_last_update_, t_current_update_ );
